--- server/api/src/main/java/vook/server/api/domain/demo/logic/DemoTermSearchService.java
@@ -1,5 +1,0 @@
-package vook.server.api.domain.demo.logic;
-
-public interface DemoTermSearchService {
-    DemoTermSearchResult search(DemoTermSearchCommand params);
-}
