+++ server/api/src/main/java/vook/server/api/web/term/usecase/RetrieveTermUseCase.java
@@ -7,6 +7,7 @@ import vook.server.api.domain.user.logic.UserLogic;
 import vook.server.api.domain.vocabulary.logic.vocabulary.VocabularyLogic;
 import vook.server.api.domain.vocabulary.model.term.Term;
 import vook.server.api.domain.vocabulary.model.vocabulary.Vocabulary;
+import vook.server.api.domain.vocabulary.service.TermQueryService;
 import vook.server.api.globalcommon.annotation.UseCase;
 import vook.server.api.policy.VocabularyPolicy;
 
@@ -20,7 +21,7 @@ public class RetrieveTermUseCase {
     private final UserLogic userLogic;
     private final VocabularyLogic vocabularyLogic;
     private final VocabularyPolicy vocabularyPolicy;
-    private final TermSearchService termSearchService;
+    private final TermQueryService termQueryService;
 
     public Result execute(Command command) {
         userLogic.validateCompletedUserByUid(command.userUid());
@@ -28,7 +29,7 @@ public class RetrieveTermUseCase {
         Vocabulary vocabulary = vocabularyLogic.getByUid(command.vocabularyUid());
         vocabularyPolicy.validateOwner(command.userUid(), vocabulary);
 
-        Page<Term> termPage = termSearchService.findAllBy(command.vocabularyUid(), command.pageable());
+        Page<Term> termPage = termQueryService.findAllBy(command.vocabularyUid(), command.pageable());
         Page<Result.Tuple> tuplePage = termPage.map(Result.Tuple::from);
         return new Result(tuplePage);
     }
@@ -61,9 +62,5 @@ public class RetrieveTermUseCase {
                 );
             }
         }
-    }
-
-    public interface TermSearchService {
-        Page<Term> findAllBy(String vocabularyUid, Pageable params);
     }
 }
