+++ server/api/src/test/java/vook/server/api/domain/vocabulary/logic/vocabulary/VocabularyLogicTest.java
@@ -11,7 +11,7 @@ import vook.server.api.domain.vocabulary.model.term.Term;
 import vook.server.api.domain.vocabulary.model.term.TermRepository;
 import vook.server.api.domain.vocabulary.model.vocabulary.UserUid;
 import vook.server.api.domain.vocabulary.model.vocabulary.Vocabulary;
-import vook.server.api.infra.search.vocabulary.MeilisearchVocabularySearchService;
+import vook.server.api.infra.search.vocabulary.MeilisearchSearchManagementService;
 import vook.server.api.testhelper.IntegrationTestBase;
 import vook.server.api.testhelper.creator.TestUserCreator;
 import vook.server.api.testhelper.creator.TestVocabularyCreator;
@@ -30,7 +30,7 @@ class VocabularyLogicTest extends IntegrationTestBase {
     @Autowired
     TestVocabularyCreator vocabularyCreator;
     @Autowired
-    MeilisearchVocabularySearchService searchService;
+    MeilisearchSearchManagementService searchService;
     @Autowired
     TermRepository termRepository;
